fix(back): Remove duplicated module.
