+++ frontend/src/app/projects/project-detail/model-overview/model-overview.component.html
@@ -11,7 +11,10 @@
 <div class="models" *ngIf="models">
   <div class="models-title">
     <h2 id="title">Models</h2>
-    <a mat-button [routerLink]="['/project', project?.slug, 'models/create']">
+    <a
+      mat-button
+      [routerLink]="['/project', project?.slug, 'models', 'create']"
+    >
       <mat-icon id="add-icon">add</mat-icon>
     </a>
   </div>
